computation done from the pkl object

+++ CONSERVATION_INITIAL_CONTACTS/plot_fraction_initial.sh
@@ -1,13 +1,19 @@
-for complex in 1PVH 3SGB 1BRS 1EMV  1GCQ 2OOB  1AK4 1AY7
+while read complex 
 do
-echo Contact_wrt_initial_$complex.txt > input.txt
-echo $complex >> input.txt
-R CMD BATCH plot_fraction_contacts.R
-mv Rplots.pdf Fraction_initial_$complex.pdf
-done
-
-R CMD BATCH plot_all_fraction_contacts.R
-mv Rplots.pdf Fraction_initial_all.pdf
-
-
-
+    if [ -e Contact_wrt_initial_$complex.txt ] 
+    then
+    echo "making individual plot for $complex"
+    echo Contact_wrt_initial_$complex.txt > input.txt
+    echo $complex >> input.txt
+    R CMD BATCH plot_fraction_contacts.R
+        if [ -e Rplots.pdf ]
+        then
+        mv Rplots.pdf Fraction_initial_$complex.pdf
+        echo "File generated successfully: Fraction_initial_$complex.pdf"
+        else
+        echo "Pb: failed to generate Fraction_initial_$complex.pdf"
+        fi
+    else 
+    echo "Contact_wrt_initial_$complex.txt was not found, done nothing"
+    fi
+done <list.txt
